requirements.txt
absl-py==1.1.0
asttokens==2.0.5
azure-core==1.24.1
azure-iot-device==2.11.0
azure-storage-blob==12.12.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
cycler==0.11.0
decorator==5.1.1
deprecation==2.1.0
executing==0.8.3
exif==1.3.5
fonttools==4.33.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
gphoto2 @ file:///home/test/HumanDetectionAndCount/gphoto2-2.3.3-cp38-cp38-linux_aarch64.whl
grpcio==1.46.3
idna==3.3
importlib-metadata==4.11.4
ipython==8.4.0
isodate==0.6.1
janus==1.0.0
jedi==0.18.1
kiwisolver==1.4.2
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline==0.1.3
meson==0.62.2
msrest==0.6.21
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
plum-py==0.8.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.3.0
rsa==4.8
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
stack-data==0.2.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.0.31.post2005241907
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
traitlets==5.2.2.post1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0